combine the load pipette logic
